requirements.txt
Flask
Flask-Limiter
GitPython
flask_sqlalchemy
swaglyrics
requests
pyjwt
